moved bookkeep delete page to be a component due to it only being called as a modal
